threads, show loading fragment while alg runs in background, currently blocked with going to recommendations fragment after alg has finished
