+++ .idea/misc.xml
@@ -19,6 +19,7 @@
         <entry key="app/src/main/res/layout/fragment_camera.xml" value="0.20416666666666666" />
         <entry key="app/src/main/res/layout/fragment_camera_result.xml" value="0.19630192502532928" />
         <entry key="app/src/main/res/layout/fragment_compose.xml" value="0.21822916666666667" />
+        <entry key="app/src/main/res/layout/fragment_loading.xml" value="0.3145833333333333" />
         <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.3328804347826087" />
         <entry key="app/src/main/res/layout/fragment_song_recommendations.xml" value="0.309405517578125" />
         <entry key="app/src/main/res/layout/fragment_spotify.xml" value="0.2513586956521739" />
