+++ app/src/main/res/layout/fragment_compose.xml
@@ -42,7 +42,7 @@
         app:layout_constraintBottom_toTopOf="@+id/btnGetColors"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/innerconstraint" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/btnGetColors"
